FIX :: 자잘한 버그 수정, 프론트 친화적으로 api 수정

+++ src/main/java/com/example/codingmom/domain/user/presentation/dto/request/CreateUserDto.java
@@ -15,7 +15,7 @@ import java.util.Collections;
 @Setter
 public class CreateUserDto {
     @NotNull
-    private Long type;
+    private int type;
     private String category;
     @NotNull
     private String kakaoid;
